Separate the routes
